feat: Centralizing types from Address components
