+++ components/custom/AddressInterface/AddNewAddress/AddNewAddress.tsx
@@ -2,27 +2,7 @@ import React, { FC } from "react";
 import { View, StyleSheet } from "react-native";
 import { AddNewAddressForm } from "./components/AddNewAddressForm";
 import { AddNewAddressViaGPS } from "./components/AddNewAddressViaGPS";
-
-export interface AddNewAddressProps {
-  latitude?: number | string;
-  longitude?: number | string;
-  postalCode: string;
-  number: string;
-  street: string;
-  complement: string;
-  neighborhood: string;
-  city: string;
-  state: string;
-  setLatitude?: (value: number) => void;
-  setLongitude?: (value: number) => void;
-  setPostalCode: (value: string) => void;
-  setNumber: (value: string) => void;
-  setStreet: (value: string) => void;
-  setComplement: (value: string) => void;
-  setNeighborhood: (value: string) => void;
-  setCity: (value: string) => void;
-  setState: (value: string) => void;
-}
+import { AddNewAddressProps } from "../types";
 
 const AddNewAddress: FC<AddNewAddressProps> = ({
   latitude,
