+++ components/custom/AddressInterface/AddNewAddress/components/AddNewAddressForm/AddNewAddressForm.tsx
@@ -1,22 +1,6 @@
 import React, { FC } from "react";
 import { View, Text, TextInput, StyleSheet } from "react-native";
-
-interface AddNewAddressFormProps {
-  postalCode: string;
-  number: string;
-  street: string;
-  complement: string;
-  neighborhood: string;
-  city: string;
-  state: string;
-  setPostalCode: (value: string) => void;
-  setNumber: (value: string) => void;
-  setStreet: (value: string) => void;
-  setComplement: (value: string) => void;
-  setNeighborhood: (value: string) => void;
-  setCity: (value: string) => void;
-  setState: (value: string) => void;
-}
+import { AddNewAddressFormProps } from "../../../types";
 
 const AddNewAddressForm: FC<AddNewAddressFormProps> = ({
   postalCode,
